Refactoring: make a couple of global enums scoped

+++ src/sqlite3gen.cpp
@@ -1128,8 +1128,8 @@ static void associateMember(const MemberDef *md, struct Refid member_refid, stru
     bindIntParameter(member_insert, ":scope_rowid", scope_refid.rowid);
     bindIntParameter(member_insert, ":memberdef_rowid", member_refid.rowid);
 
-    bindIntParameter(member_insert, ":prot", md->protection());
-    bindIntParameter(member_insert, ":virt", md->virtualness());
+    bindIntParameter(member_insert, ":prot", static_cast<int>(md->protection()));
+    bindIntParameter(member_insert, ":virt", static_cast<int>(md->virtualness()));
     step(member_insert);
   }
 }
@@ -1630,7 +1630,7 @@ static void generateSqlite3ForMember(const MemberDef *md, struct Refid scope_ref
 
   bindIntParameter(memberdef_insert,":rowid", refid.rowid);
   bindTextParameter(memberdef_insert,":kind",md->memberTypeName());
-  bindIntParameter(memberdef_insert,":prot",md->protection());
+  bindIntParameter(memberdef_insert,":prot",static_cast<int>(md->protection()));
 
   bindIntParameter(memberdef_insert,":static",md->isStatic());
   bindIntParameter(memberdef_insert,":extern",md->isExternal());
@@ -1662,7 +1662,7 @@ static void generateSqlite3ForMember(const MemberDef *md, struct Refid scope_ref
     bindIntParameter(memberdef_insert,":optional",md->isOptional());
     bindIntParameter(memberdef_insert,":required",md->isRequired());
 
-    bindIntParameter(memberdef_insert,":virt",md->virtualness());
+    bindIntParameter(memberdef_insert,":virt",static_cast<int>(md->virtualness()));
   }
 
   if (md->memberType() == MemberType_Variable)
@@ -1922,7 +1922,7 @@ static void generateSqlite3ForClass(const ClassDef *cd)
   bindTextParameter(compounddef_insert,":name",cd->name());
   bindTextParameter(compounddef_insert,":title",cd->title());
   bindTextParameter(compounddef_insert,":kind",cd->compoundTypeString());
-  bindIntParameter(compounddef_insert,":prot",cd->protection());
+  bindIntParameter(compounddef_insert,":prot",static_cast<int>(cd->protection()));
 
   int file_id = insertPath(cd->getDefFileName());
   bindIntParameter(compounddef_insert,":file_id",file_id);
@@ -1991,8 +1991,8 @@ static void generateSqlite3ForClass(const ClassDef *cd)
     struct Refid derived_refid = insertRefid(cd->getOutputFileBase());
     bindIntParameter(compoundref_insert,":base_rowid", base_refid.rowid);
     bindIntParameter(compoundref_insert,":derived_rowid", derived_refid.rowid);
-    bindIntParameter(compoundref_insert,":prot",bcd.prot);
-    bindIntParameter(compoundref_insert,":virt",bcd.virt);
+    bindIntParameter(compoundref_insert,":prot",static_cast<int>(bcd.prot));
+    bindIntParameter(compoundref_insert,":virt",static_cast<int>(bcd.virt));
     step(compoundref_insert);
   }
 
@@ -2003,8 +2003,8 @@ static void generateSqlite3ForClass(const ClassDef *cd)
     struct Refid base_refid = insertRefid(cd->getOutputFileBase());
     bindIntParameter(compoundref_insert,":base_rowid", base_refid.rowid);
     bindIntParameter(compoundref_insert,":derived_rowid", derived_refid.rowid);
-    bindIntParameter(compoundref_insert,":prot",bcd.prot);
-    bindIntParameter(compoundref_insert,":virt",bcd.virt);
+    bindIntParameter(compoundref_insert,":prot",static_cast<int>(bcd.prot));
+    bindIntParameter(compoundref_insert,":virt",static_cast<int>(bcd.virt));
     step(compoundref_insert);
   }
 
@@ -2305,7 +2305,7 @@ static void generateSqlite3ForGroup(const GroupDef *gd)
   // + classes
   writeInnerClasses(gd->getClasses(),refid);
 
-  // + concepts 
+  // + concepts
   writeInnerConcepts(gd->getConcepts(),refid);
 
   // + namespaces
